problem_1/requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==2.0.9
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
openapi-codec==1.3.2
pytz==2018.7
PyYAML==3.13
requests==2.20.1
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.24.1
